Rework source mapping for DCSS messages
Instead of designating slot numbers for each source and mapping those to
source information, give each query an id and a map to the source
information. Also record the time of the query and ignore results past
an 80 second timeout, reclaiming their query id also using that same
timeout. The use of query ids and the timeout prevents messages from
getting misrouted to the wrong source under some conditions. Previously
when a game ended and a new one is watched and got the old games slot
number, it was possible for messages from the old game to get routed to
the new one. With this commit slot numbers for a connection are no
longer a thing.

This also puts a limit of the number of outstanding queries we'll allow
at once. The current limit is 100, but it can be increased if it's found
to be too low.