Fixes and refactors for chat limits, config loading, and single user
Fix the chat limits so they actually work.

Admins now only get the ability to run commands for other users and have
no special status where they're always watched or they get to ignore
chat command limits.

Fix config loading so that the right fields are checked and refactor the
loading code. The webtiles.servers table is gone; instead a websocket
url should be directly given in server_url. There's more consistency in
requiring fields only if features that depend on them are enabled. The
sample config file is updated accordingly and with better documentation.

Now single user mode is better supported. It requires listen_user to be
defined for any enabled service and updates the database so that this
user actually exists and has their WebTiles and Twitch usernames linked
if both services are enabled. beem will now not accept any commands in
single user mode in either service aside from nick and twitch-reminder
(for WebTiles). Under single user mode, if Twitch is enabled, beem
listens to the chat of listen_user at startup instead of listening to
the channel of the bot itself.