Trim and update the command guide
This version reduces the amount of extra detail that should instead go
into LearnDB or !help documentation. The guide is easier to quickly
browse to see examples of commands you might like to run, with
references to the apporpriate LearnDB entries for further details and
examples. The overall amount of text is reduced and clarified, and the
command examples shown are more relevant.