Properly shut down connections when a game ends
We were logging these events and passing them onto the base
handle_message of WebTilesGameConnection, but this did not actually stop
the connection task. Hence connections were persisting in the set of
subscriber connections. This possibly accounts for the weird behaviour
seen where beem wasn't joining some games when it should, mostly likely
because the the connectings were persisting until the
max_watched_subscribers limit was reached.