Improve help text lookup
The bot only prints its help message once (and only when autowatching),
so this commit makes it easier for users trying to find the help
text. I've made the learnDB text for `??beem` have an appropriate
version of this text and added a `!beem` command to Sequell which looks
this up. The `!help` text for this command also tells people how to look
up help. Now the various things I see people trying in chat, like
`!beem' with no arguments, `!help beem' and `!help !beem' etc will all
get them to the help text.