Split off Twitch support to its own package and generalize code
The Twitch side of beem is being split into its own python package that
depends on this one. The bot previously used the Twitch account 'r4nr',
but will now use the account 'LomLobot' and the new project takes this
name as well:

https://github.com/gammafunk/lomlobot

This commit generalizes the ChatWatcher, DCSSManager, and adds Config
and UserDB classes, and those first two no longer depend on
beem-specific configuration. The manager classes also no longer need to
be instantiated when their modules are loaded.

The BeemConfig class now inherents from a base Config class, which
handles the TOML parsing and configuration common to both beem and
LomLobot. The user DB code has been generalized and uses the class
UserDB to represent a single user table that loads/stores the sqlite3
data.

LomLobot has its own repository, so code to specific to that project has
been deleted from this one. It will continue to depend on beem for the
forseeable future.

This commit also adds a GPL2 license file.