Do periodic WebTiles connection checks and refactor
A ongoing problem with beem has been websocket connections that
eventually become dead in such a way the connection's read would
continue forever without receiving data. This would only be seen through
exceptions raised when the game actually became idle as seen through the
lobby entry, leading to the manager trying to shut down the connection,
or when beem itself was shut down. In either case this would raise an
internal, logged exception in the websocket module and generate a
connection-related exception when calling the websockets close(). The
log message for the internal error was:

Failing the WebSocket connection: 1011

but it's not clear why this was happening. The issue was serious since
beem's connection to games it was supposed to be watching was dead, but
it couldn't know this since no exception was being raised in the start()
loop.

To help address the issue, this commit adds a task that periodically
sends a websocket ping every 60 seconds. This task is created by the
connections start() task and is canceled by the webtiles manager along
with the start() task when shutting down the connection. If a ping fails
to send, an exception will hopefully get generated, although this may
run into the same problem as seen with recv() where it waits forever. In
preliminary testing it seems to prevent generation of these failing
websockets in the first place, which may be enough.

This commit also refactors the connection classes to use a
ConnectionHandler class that has code related to the ping and start()
loops, that latter being mostly duplicated between the two classes
before this commit. The exception logging has also been moved to a
function that handles the log message setup, which was also duplicated
in many places, and this refactor has also been done in the DCSS
manager.