Add SSL and SASL authentication support for IRC
The use_ssl config option enables use of SSL (either with or without
SASL) and setting the password option will have the server use SASL
authentication (without requiring SSL).

This extends some of the classes in the irc package to add basic
capability requests and to handle sasl authentication. Setting the
password argument to ServerConnection.connect() will have the server
automatically request sasl capabilities and then negotiate
authentication with the server. The authenticated property will be true
once SASL authentication completes.

On the DCSS manager side, the loggen_in property is replaced with
ready() to check that the server is ready to respond to
requests.