Allow setting an IRC username for SASL authentication
Previously we used the nick, but the nick may be different from the
login credentials, especially in the case of beem where we now have a
nick for each server that's grouped with the main account nick.

This commit also improves and refactors config checking to handle cases
when we require certain other fields only if a specific field is
defined.