Add a !beem player-only command
Add a new player_only field to the user DB which can be set to on/off by
this command. When on, beem only responds to the player's DCSS commands.
A few people have requested this functionality to minimize the amount of
bot spam in their channels while still giving themselves access. It
could be expanded to allow the player to set 'moderator' users, but this
is probably unnecessary.

The beem control commands work for everyone when player_only is set,
although this could be changed to also disallow these or perhaps all but
e.g. !beem subscribe.