Fix handling for DCSS results with no query
For DCSS query results, we tried to look up the query ID found in
the result without checking that the dict of queries had this key. The
entry can be removed if the query times out yet a result somehow still
comes from Sequell. Properly check that the key exists and return None
when it doesn't.