Make !help an alias to !bothelp
This helps avoid confusion, since most users are looking for this info
instead of the info in Sequell's help. There's now a !sequellhelp
command in Sequell that's an alias to !help, so this can still be
accessed. The command guide and probably the bot !help command itself
should be updated to mention this.