Generalize the DB to allow multiple tables
Previously the DB code was centered around there only being a single
user table. This commit generalizes the database code to allow arbitrary
tables each of which can have multiple primary keys. Table definitions
are per-manager as before, since e.g. Discord has different needs from
WebTiles. The DB data is in a memory cache using tuples of primary key
fields as keys in a dict for each table. All tables in the table
specification are checked and created as necessary at runtime.

Existing databases should be modified so that, in the webtiles_users
table, username is the primary key and the id field is removed.
Databases in the old format will continue to work normally for now, but
future changes might rely on having a correct primary key. Anyone
upgrading is advised to recreate that table and import the existing rows
using the sqlite3 command.

The DB code is now in the source file botdb.py instead of userdb.py.