Improved exception logging
Now we log the traceback in addition to the exception type and value.
Before it could be very difficult to know what line and code path
generated an exception for some kinds of errors.

A few kinds of routine errors need more specific handling to ommit the
traceback spam, but these would most notably be basic network connection
errors.