Move query tracking out of the DCSSManager and into the IRCBot class
Previously DCSSManager coordinated data for all queries through a single
dict and had each bot instance maintain its own queue to track
outstanding queries for the synchronous bots. Tracking query data in
DCSSManager isn't necessary, since query data doesn't need to be shared
between bots, and moving all tracking to the bot instances makes
tracking the type of query from when it was received simpler to
implement. We want to add info on the type of query (e.g. Sequell vs
monster vs git) to the query tracking instead of matching the resulting
message, since the latter is pattern-based for synchronous bots and
hence error-prone.

This commit moves all query tracking logic into IRCbot, so each bot has
its own set of query slots. This slot sized is 62, one for each
character in a-zA-Z0-9, since we now use one of these characters as
prefixes for Sequell instead of two digits. This frees up an extra
character for Sequell results.

Additionally the bot configuration looks for a regex pattern field for
each of the three supported bot services, Sequell, monster, and git,
with each containing a list of patterns to match only that type of
service. This distinction lets beem know the kind of query being made at
the time of query, which is necessary for routing the result. The
specific fields it looks for are named after the service:
'sequell_patterns', 'monster_patterns', and 'git_patterns'. As before, a
bot running Sequell can't support the other two services, but a single
bot can support both of monster and git.

This commit fixes a long-standing bug seen in beem and Cerebot where
query results could get directed to the wrong originating chat source if
queries came in from two different sources before the timeout period.