Reconnect upon SASL authentication failure instead of shutting down
Previously we killed the process if SASL authentication failed since it
was assumed that this only happens if an invalid password is configured.
However Freenode sometimes loses connectivity and subsequently fails
SASL authentication for a while after the connection is restored. Since
we don't want the server to shut down in this case, this commit has the
DCSSManager but to continue trying to connecting and authenticating.